AmpereAltraPkg: Fix SMBUS not generating STOP condition
Currently, we don't read any further bytes when the SMBUS received
count is 0 and the PEC check is disabled instead trying to complete
the transaction. Due to the lack of a STOP condition, it is incorrect.
The same thing happens when the count exceeds the buffer length.

Because we cannot trigger the STOP condition the first time we read
the count byte, we must read another dummy byte to obtain
the STOP condition.

Signed-off-by: Tinh Nguyen <tinhn@amperecomputing.com>