AmpereAltraPkg: Fix OS kernel crash if not call SetVirtualAddressMap()
OS kernel avoided SetVirtualAddressMap() on arm64 due to commit d3549a93
("efi/arm64: libstub: avoid SetVirtualAddressMap() when possible").
Currently, RuntimeFlashLib in AmpereAltraPkg relies on function call
SetVirtualAddressMap() to determine whether it is in boot time or
run-time. At boot time, the code optimizes to avoid the use of temporary
buffer because the virtual address and physical address are identical.
If SetVirtualAddressMap() isn't called, it does not know when to switch.
As such, don't do that optimization and the physical address is always
known.

Signed-off-by: Chuong Tran <chuong@os.amperecomputing.com>