Set correct CpuStatus value in OemGetProcessorInformation
The IsCpuEnabled function returns whether the specified CPU _core_ is
enabled, while OemGetProcessorInformation is asking about an entire
processor. So, fix the code always set the primary socket as enabled,
and the secondary if IsSlaveSocketActive returns TRUE.

Also, if `OemIsProcessorPresent (ProcessorIndex)` returns FALSE, it
likely indicates that it's asking about the secondary processor so its
return value just means the processor isn't present - so we can return
TRUE from OemGetProcessorInformation.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>