Silicon/Ampere: Revert CpuConfigDxe change that removed EFI variables
In commit 3357b7d72b154fadab1d1c445014c3a6394ac6a3, the usage of EFI
variables was removed. However that was wrong since the configuration of
whether to expose the SLC as an L3 cache in the PPTT isn't stored as an
NVPARAM but only in the EFI variable. So, partially revert that commit
to re-add the EFI variables.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>